[iac] Consolidate the reviewed import command
